Re-organize animation importer
In particular, remember all the action each one creates (we're
going to put them in NLA strips).
